requirements.txt
fastapi
pillow